Code cleanup, remove deprecated methods
In particular, remove legacy detection classifiers and related code. They are much less maintainable than the current ObjectClassifier implementation, and relied upon serialization + retraining.

Serialized classifiers from older QuPath releases couldn't even be read in the latest version due to serialization incompatibilities, or could sometimes be read but would not give the expected results (e.g. because of differences in feature calculations).

One significant change is that `PathClassifierTools` has been removed, with the relevant methods moved primarily to `PathObjectTools` or (for channels) `ServerTools`.
