+++ qupath-core-processing/src/main/java/qupath/lib/plugins/objects/SmoothFeaturesPlugin.java
@@ -33,7 +33,6 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import qupath.lib.classifiers.PathClassifierTools;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.servers.ImageServer;
@@ -149,7 +148,7 @@ public class SmoothFeaturesPlugin<T> extends AbstractInteractivePlugin<T> {
 						return;
 					
 					// TODO: ACCESS & USE THE CLASSIFIER DATA!!!!
-					List<String> measurements = new ArrayList<>(PathClassifierTools.getAvailableFeatures(pathObjects));
+					List<String> measurements = new ArrayList<>(PathObjectTools.getAvailableFeatures(pathObjects));
 					Iterator<String> iter = measurements.iterator();
 					while (iter.hasNext()) {
 						String name = iter.next().toLowerCase();
