--- qupath-core-processing/src/main/java/qupath/opencv/classify/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Classifiers using OpenCV's ml module.
- */
-package qupath.opencv.classify;
