+++ qupath-core/src/main/java/qupath/lib/classifiers/package-info.java
@@ -1,10 +1,7 @@
 /**
  * Classes and interfaces to support machine learning classifiers.
  * <p>
- * Most classes directly in this package are scheduled for removal in the next QuPath release, 
- * this these refer to specific object classifier implementations that are no longer the default since 
- * QuPath v0.2.
- * Most of their functionality has been replaced by the {@link qupath.lib.classifiers.object} subpackage.
- * <p>
+ * This was the main package in QuPath v0.1.2 and earlier.
+ * Since v0.2.0, the new subpackages have been created for pixel and object classifiers.
  */
 package qupath.lib.classifiers;
