+++ qupath-core/src/test/java/qupath/lib/color/TestColorsToolsAwt.java
@@ -22,14 +22,8 @@
 package qupath.lib.color;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.util.Locale;
-
 import org.junit.jupiter.api.Test;
 
-import qupath.lib.color.ColorTransformer.ColorTransformMethod;
 import qupath.lib.common.ColorTools;
 
 @SuppressWarnings("javadoc")
