--- qupath-extension-processing/src/main/java/qupath/process/gui/ml/legacy/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Legacy classes related to object classification.
- * These how now been mostly superseded and are likely to be removed in a later version.
- */
-@Deprecated
-package qupath.process.gui.ml.legacy;
